# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dwizzyBRAIN API",
    "version": "1.8.0",
    "description": "Standardized read API for market snapshots, OHLCV candles, aggregated tickers, arbitrage signals, quant pattern search, DeFi registry reads, news feeds, and auth/session endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/v1/market": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "List market snapshots",
        "operationId": "listMarketSnapshots",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated market snapshot list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/market/{id}": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Get market snapshot detail",
        "operationId": "getMarketSnapshot",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          }
        ],
        "responses": {
          "200": {
            "description": "Market snapshot detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDetailResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/market/{id}/ohlcv": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Get OHLCV candles",
        "operationId": "getMarketOHLCV",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          },
          {
            "$ref": "#/components/parameters/Exchange"
          },
          {
            "$ref": "#/components/parameters/Timeframe"
          },
          {
            "$ref": "#/components/parameters/OHLCVLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "OHLCV candles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OHLCVResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/market/{id}/tickers": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Get aggregated ticker snapshot",
        "operationId": "getMarketTickers",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated ticker snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/market/{id}/orderbook": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Get top-of-book snapshot",
        "operationId": "getMarketOrderBook",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          }
        ],
        "responses": {
          "200": {
            "description": "Top-of-book snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/market/{id}/arbitrage": {
      "get": {
        "tags": [
          "Market"
        ],
        "summary": "Get recent arbitrage signals",
        "operationId": "getMarketArbitrage",
        "parameters": [
          {
            "$ref": "#/components/parameters/CoinID"
          },
          {
            "$ref": "#/components/parameters/ArbitrageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Recent arbitrage signals",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArbitrageResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/defi": {
      "get": {
        "tags": [
          "DeFi"
        ],
        "summary": "Get DeFi overview",
        "operationId": "getDefiOverview",
        "parameters": [
          {
            "$ref": "#/components/parameters/OverviewLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "DeFi overview payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefiOverviewResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/defi/protocols": {
      "get": {
        "tags": [
          "DeFi"
        ],
        "summary": "List DeFi protocols",
        "operationId": "listDefiProtocols",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "$ref": "#/components/parameters/DefiCategory"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated DeFi protocol list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefiProtocolListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/defi/protocols/{slug}": {
      "get": {
        "tags": [
          "DeFi"
        ],
        "summary": "Get DeFi protocol detail",
        "operationId": "getDefiProtocol",
        "parameters": [
          {
            "$ref": "#/components/parameters/DefiSlug"
          }
        ],
        "responses": {
          "200": {
            "description": "DeFi protocol detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefiProtocolResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/defi/chains": {
      "get": {
        "tags": [
          "DeFi"
        ],
        "summary": "List DeFi chains",
        "operationId": "listDefiChains",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Chain TVL list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefiChainListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/v1/defi/dexes": {
      "get": {
        "tags": [
          "DeFi"
        ],
        "summary": "List DeFi DEXes",
        "operationId": "listDefiDexes",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "DEX volume list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefiDexListResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "CoinID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Canonical coin identifier such as bitcoin or ethereum.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of items to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 20
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "description": "Zero-based offset into the collection.",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "Exchange": {
        "name": "exchange",
        "in": "query",
        "required": false,
        "description": "Exchange identifier such as binance, bybit, kraken, okx, kucoin, gateio, mexc, or htx.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "Timeframe": {
        "name": "timeframe",
        "in": "query",
        "required": false,
        "description": "OHLCV timeframe.",
        "schema": {
          "type": "string",
          "enum": [
            "1m",
            "5m",
            "15m",
            "1h",
            "4h",
            "1d",
            "1w"
          ],
          "default": "1m"
        }
      },
      "OHLCVLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of candles to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 200
        }
      },
      "OverviewLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of items to show in each overview section.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 20,
          "default": 5
        }
      },
      "DefiSlug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "DeFi protocol slug.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "DefiCategory": {
        "name": "category",
        "in": "query",
        "required": false,
        "description": "Optional protocol category filter.",
        "schema": {
          "type": "string"
        }
      },
      "ArbitrageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of arbitrage signals to return.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ForbiddenError": {
        "description": "Authenticated but insufficient plan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0,
            "example": 1000
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200,
            "example": 20
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "example": 0
          }
        }
      },
      "MarketListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "ArbitrageSignal": {
        "type": "object",
        "required": [
          "id",
          "detected_at",
          "coin_id",
          "symbol",
          "buy_exchange",
          "sell_exchange",
          "buy_price",
          "sell_price",
          "gross_spread_pct",
          "is_profitable",
          "buy_depth_usd",
          "sell_depth_usd",
          "alerted"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "detected_at": {
            "type": "string",
            "format": "date-time"
          },
          "coin_id": {
            "type": "string",
            "example": "bitcoin"
          },
          "symbol": {
            "type": "string",
            "example": "BTC/USDT"
          },
          "buy_exchange": {
            "type": "string",
            "example": "gateio"
          },
          "sell_exchange": {
            "type": "string",
            "example": "htx"
          },
          "buy_price": {
            "type": "number"
          },
          "sell_price": {
            "type": "number"
          },
          "gross_spread_pct": {
            "type": "number"
          },
          "is_profitable": {
            "type": "boolean"
          },
          "buy_depth_usd": {
            "type": "number"
          },
          "sell_depth_usd": {
            "type": "number"
          },
          "alerted": {
            "type": "boolean"
          }
        }
      },
      "SnapshotSummary": {
        "type": "object",
        "required": [
          "coin_id",
          "symbol",
          "name",
          "rank",
          "updated_at",
          "availability",
          "exchange_count"
        ],
        "properties": {
          "coin_id": {
            "type": "string",
            "example": "bitcoin"
          },
          "symbol": {
            "type": "string",
            "example": "btc"
          },
          "name": {
            "type": "string",
            "example": "Bitcoin"
          },
          "image_url": {
            "type": "string",
            "format": "uri"
          },
          "rank": {
            "type": "integer",
            "example": 1
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "exchange_count": {
            "type": "integer",
            "minimum": 0
          },
          "current_price_usd": {
            "type": "number"
          },
          "price_source": {
            "$ref": "#/components/schemas/PriceSource"
          }
        }
      },
      "SnapshotDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SnapshotSummary"
          },
          {
            "type": "object",
            "required": [
              "cold_data",
              "mappings",
              "exchange_prices"
            ],
            "properties": {
              "cold_data": {
                "$ref": "#/components/schemas/ColdData"
              },
              "mappings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MappingSnapshot"
                }
              },
              "exchange_prices": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ExchangePrice"
                }
              },
              "recent_arbitrage": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ArbitrageSignal"
                }
              }
            }
          }
        ]
      },
      "MarketDetailResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SnapshotDetail"
          }
        }
      },
      "OHLCVPoint": {
        "type": "object",
        "required": [
          "timestamp",
          "coin_id",
          "exchange",
          "symbol",
          "timeframe",
          "open",
          "high",
          "low",
          "close",
          "volume",
          "quote_volume",
          "trades",
          "is_closed"
        ],
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "coin_id": {
            "type": "string",
            "example": "bitcoin"
          },
          "exchange": {
            "type": "string",
            "example": "binance"
          },
          "symbol": {
            "type": "string",
            "example": "BTC/USDT"
          },
          "timeframe": {
            "type": "string",
            "example": "1m"
          },
          "open": {
            "type": "number"
          },
          "high": {
            "type": "number"
          },
          "low": {
            "type": "number"
          },
          "close": {
            "type": "number"
          },
          "volume": {
            "type": "number"
          },
          "quote_volume": {
            "type": "number"
          },
          "trades": {
            "type": "integer"
          },
          "is_closed": {
            "type": "boolean"
          }
        }
      },
      "OHLCVResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OHLCVPoint"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "coin_id",
              "exchange",
              "timeframe",
              "limit"
            ],
            "properties": {
              "coin_id": {
                "type": "string",
                "example": "bitcoin"
              },
              "exchange": {
                "type": "string",
                "example": "binance"
              },
              "timeframe": {
                "type": "string",
                "example": "1m"
              },
              "limit": {
                "type": "integer",
                "minimum": 1
              }
            }
          }
        }
      },
...
      "TickerSnapshot": {
        "type": "object",
        "required": [
          "coin_id",
          "exchange_count",
          "exchanges"
        ],
        "properties": {
          "coin_id": {
            "type": "string",
            "example": "bitcoin"
          },
          "best_bid": {
            "type": "number"
          },
          "best_bid_exchange": {
            "type": "string",
            "example": "htx"
          },
          "best_ask": {
            "type": "number"
          },
          "best_ask_exchange": {
            "type": "string",
            "example": "kraken"
          },
          "cross_exchange_spread_pct": {
            "type": "number"
          },
          "exchange_count": {
            "type": "integer",
            "minimum": 0
          },
          "exchanges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TickerExchange"
            }
          }
        }
      },
      "TickerResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TickerSnapshot"
          }
        }
      },
      "ArbitrageResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArbitrageSignal"
            }
          }
        }
      },
...
      "DefiChainSummary": {
        "type": "object",
        "required": [
          "chain",
          "updated_at"
        ],
        "properties": {
          "chain": {
            "type": "string",
            "example": "ethereum"
          },
          "tvl_usd": {
            "type": "number"
          },
          "tvl_change_1d_pct": {
            "type": "number"
          },
          "tvl_change_7d_pct": {
            "type": "number"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DefiOverview": {
        "type": "object",
        "required": [
          "protocols",
          "chains",
          "dexes"
        ],
        "properties": {
          "protocols": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiProtocolSummary"
            }
          },
          "chains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiChainSummary"
            }
          },
          "dexes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiDexSummary"
            }
          }
        }
      },
      "DefiOverviewResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DefiOverview"
          },
          "meta": {
            "type": "object",
            "required": [
              "limit"
            ],
            "properties": {
              "limit": {
                "type": "integer",
                "minimum": 1
              }
            }
          }
        }
      },
      "DefiProtocolListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiProtocolSummary"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "total",
              "limit",
              "offset"
            ],
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0
              },
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 200
              },
              "offset": {
                "type": "integer",
                "minimum": 0
              },
              "category": {
                "type": "string"
              }
            }
          }
        }
      },
      "DefiProtocolResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DefiProtocolSummary"
          }
        }
      },
      "DefiChainListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiChainSummary"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "limit"
            ],
            "properties": {
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 200
              }
            }
          }
        }
      },
      "DefiDexListResponse": {
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DefiDexSummary"
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "limit"
            ],
            "properties": {
              "limit": {
                "type": "integer",
                "minimum": 1,
                "maximum": 200
              }
            }
          }
        }
      },
...
    }
  }
}